+++ script/__init__.py
@@ -1,8 +1,0 @@
-from .theme_config import ThemeConfig
-from .package_config import PackageConfig
-from .theme_builder import compile_theme
-from .installer import install_theme, remove_theme
-from .active_theme import active_theme
-from .packaging import package_debian
-from .dfile import diff_file
-from .man_builder import compile_manpage